requirements.txt
streamlit>=1.25.0
openai>=0.27.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=0.21.0
python-binance>=0.7.9
pandas-ta>=0.3.14b0
rpds-py>=0.10.0
cryptography>=39.0.0
aiohttp>=3.8.0
websockets>=10.0
dateparser>=1.1.0
pytz>=2022.1
tzlocal>=4.2
ujson>=5.4.0
gitpython>=3.1.30
matplotlib>=3.5.0